Read receipts as UUIDs.
This makes correlating a application's log with the broker's stomp log a lot
easier. Consider the scenario where a web application is instantiating a Client
object per web request, sends a couple of messages then closes it.  There'll be
thousands of entries in the broker's stomp log where all messages and receipts
have id values that rarely exceeded 2 or 3, say.

(ppaul/6540400, #5)